requirements.txt
python-decouple
beautifulsoup4
selenium
Pillow
numpy
pandas
